requirements.txt
google-generativeai>=0.3.2
pymilvus>=2.3.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
pydantic>=2.0.0
requests>=2.31.0
aiohttp>=3.8.0
streamlit>=1.28.0
supabase>=2.0.0
neo4j>=5.0.0
pathlib2>=2.3.0
typing-extensions>=4.5.0 